Add an example to run postgresql locally with docker for miniflux2
The code snippet allow to easily run a local Postgresql
database to test miniflux. Destroying and re-creating the
database is easy and allow for easy testing. Commands are
pre-configured for miniflux2 either for integration-tests or
manual tests. (based on default configuration of miniflux2)

I add it here so I and futur developer can easily reproduce it.
It's not complicated to build from the docker hub website, but
make things quicker.